Split config.rs into loading, raw types, and validation modules (#134)
